修改部分类型问题，合并level和chapter目录
